# openapi.yaml
openapi: 3.0.3
info:
  title: Codex Usage Dashboard API
  version: 1.0.0
  description: |
    当前项目的本地管理 API。服务端使用配置的 ChatGPT OAuth Bearer Token
    读取 wham/usage 和统计接口，前端或其他客户端只需要调用本项目的
    /api/* 接口，不需要直接接触 OAuth 凭证。

    这不是 OpenAI 官方 API，也不是聊天请求代理；它只提供额度、统计、预测、
    配置和提示音资源。

    所有 HTTP 响应都会通过 X-Codex-Meter-Version、X-Codex-Meter-Commit
    和 X-Codex-Meter-Build-Time 响应头提供当前构建标识。
servers:
  - url: http://127.0.0.1:8123
    description: 本地服务示例
  - url: https://codex.your-domain.invalid/codex
    description: 通过反向代理发布时的示例地址，请替换域名和路径
...
paths:
...
  /api/usage:
    get:
      tags: [usage]
      operationId: getUsage
      summary: 获取当前额度
      description: |
        account_id 为本项目账号 id；省略或空字符串时固定读取主账号，与页面 active_account_id 无关。
        旧调用保留原响应字段；只有显式指定 account_id 的响应包含 account_id、account_name。
        每个账号的凭证、代理、缓存和历史独立。
        服务端读取 ChatGPT 的 wham/usage，并转换为前端使用的稳定结构。
        force=true 会绕过本地缓存并立即请求上游；省略 force 或设为 false
        时使用服务端缓存策略。
      security:
        - basicAuth: []
        - appApiKey: []
        - appBearer: []
      parameters:
        - $ref: '#/components/parameters/AccountID'
        - $ref: '#/components/parameters/Force'
      responses:
        '200':
          description: 当前额度
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/AccountConflict'
        '502':
          $ref: '#/components/responses/UpstreamError'
  /api/usage/analytics:
    get:
      tags: [usage]
      operationId: getUsageAnalytics
      summary: 获取日期范围使用统计
      description: |
        account_id 为本项目账号 id；省略或空字符串时固定读取主账号，与页面 active_account_id 无关。
        旧调用保留原响应字段；只有显式指定 account_id 的响应包含 account_id、account_name。
        每个账号的凭证、代理、缓存和历史独立。
        服务端聚合指定日期范围的 daily-token-usage-breakdown 和
        daily-workspace-usage-counts 数据，返回每日 token、额度、工作区统计以及按模型的使用占比。
        未提供日期范围时兼容返回最近七个已完成日期（不包含今天）；上游统计不可用时返回 502。
      security:
        - basicAuth: []
        - appApiKey: []
        - appBearer: []
      parameters:
        - $ref: '#/components/parameters/AccountID'
        - $ref: '#/components/parameters/AnalyticsStartDate'
        - $ref: '#/components/parameters/AnalyticsEndDate'
        - $ref: '#/components/parameters/Force'
      responses:
        '200':
          description: 指定日期范围统计
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageAnalytics'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/AccountConflict'
        '502':
          $ref: '#/components/responses/UpstreamError'
  /api/accounts:
    get:
      tags: [accounts]
      operationId: listAccounts
      summary: 获取脱敏账号列表
      description: 返回账号名称、凭证配置状态和默认账号；不请求上游，不返回 OAuth Token 或 Cookie 原文。
      security:
        - basicAuth: []
        - appApiKey: []
        - appBearer: []
      responses:
        '200':
          description: 账号列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountsView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [accounts]
      operationId: createAccount
      summary: 新增监控账号
      description: 使用独立凭证新增账号（最多 32 个），服务端生成本地 id。需要 access_token 和 chatgpt_account_id；首次配置时自动设为页面展示账号。未提供请求上下文时使用默认值，代理默认直连，不继承当前账号的凭证或代理。保存不触发上游请求，可先调用配置测试接口。
      security:
        - basicAuth: []
        - appApiKey: []
        - appBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountUpdate'
      responses:
        '201':
          description: 新增成功，返回完整脱敏列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountsView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/accounts/usage:
    get:
      tags: [accounts]
      operationId: getAccountsUsage
      summary: 获取所有账号额度
      description: 最多并发查询 4 个账号，返回各自额度、缓存状态和错误。单个账号失败不会影响其他账号，整体仍返回 200；存在最近快照时返回 usage 并标记 stale=true 与 error。后台每 5 分钟分别采样所有已配置账号。
      security:
        - basicAuth: []
        - appApiKey: []
        - appBearer: []
      parameters:
        - $ref: '#/components/parameters/Force'
      responses:
        '200':
          description: 多账号额度与逐账号状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountsUsageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/accounts/active:
    put:
      tags: [accounts]
      operationId: switchActiveAccount
      summary: 切换页面展示账号
      description: 保存 active_account_id 并立即更新页面展示。页面显式传 account_id 读取所选账号；未传账号的旧应用 API 始终使用主账号。切换不会更换主账号凭证、代理、缓存或公共预测缓存。所有账号仍独立采样，不修改 Codex CLI 登录。
      security:
        - basicAuth: []
        - appApiKey: []
        - appBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountSwitch'
      responses:
        '200':
          description: 已切换，返回完整脱敏列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountsView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/AccountConflict'
  /api/accounts/{account_id}:
    parameters:
      - name: account_id
        in: path
        required: true
        description: 本项目账号 id，从账号列表中获取。
        schema:
          type: string
    put:
      tags: [accounts]
      operationId: updateAccount
      summary: 更新账号名称或凭证
      description: 部分更新指定账号，未提供字段保持不变。修改凭证仅使该账号缓存失效；chatgpt_account_id 改变时须新增账号以保留独立历史。代理及其认证属于指定账号；修改代理只使该账号缓存失效。缓存时间和管理认证为全局设置。
      security:
        - basicAuth: []
        - appApiKey: []
        - appBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountUpdate'
      responses:
        '200':
          description: 已更新，返回完整脱敏列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountsView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [accounts]
      operationId: deleteAccount
      summary: 删除监控账号
      description: 移除账号凭证和运行时缓存，保留数据库历史归档。删除当前账号后默认切换到列表中第一个账号；禁止删除主账号或最后一个账号，返回 409。
      security:
        - basicAuth: []
        - appApiKey: []
        - appBearer: []
      responses:
        '200':
          description: 已删除，返回剩余脱敏列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountsView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/AccountConflict'
  /api/prediction:
    get:
      tags: [prediction]
      operationId: getResetPrediction
      summary: 获取重置预测
      description: |
        从 codex-resets.com 的公开状态 API 读取最新已执行重置、scheduled_reset（已公告待执行重置）、active_watch 和统计数据。
        已排期重置没有预测概率或社区投票；即使 scheduled_for 已过，也必须等待上游确认执行。
        未排期时补充读取首页社区投票。历史记录或首页暂时不可用时仍返回状态 API 的信号数据。
        该数据是公共预测信息，不代表当前账户已经完成额度重置。
      security:
        - basicAuth: []
        - appApiKey: []
        - appBearer: []
      parameters:
        - $ref: '#/components/parameters/Force'
      responses:
        '200':
          description: 重置预测数据
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResetPrediction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/UpstreamError'
  /api/config:
    get:
      tags: [config]
      operationId: getConfig
      summary: 获取非敏感配置
      description: |
        返回当前运行配置。OAuth Token 不会原文返回，只返回 token_configured
        和 token_hint。
      security:
        - basicAuth: []
        - appApiKey: []
        - appBearer: []
      responses:
        '200':
          description: 非敏感配置
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigView'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags: [config]
      operationId: updateConfig
      summary: 更新运行配置
      description: |
        部分更新配置。未提供的字段保持原值。access_token 和
        chatgpt_account_id 必须同时有效，否则服务端拒绝保存。cache_ttl 使用
        Go duration 格式，例如 10m、30s 或 0。代理与凭证按 account_id 更新，
        省略时固定使用主账号；代理用户名和密码通过 proxy_username、proxy_password 设置。
        省略密码保留已有值，proxy_clear_auth=true 清除认证。缓存时间和管理认证为全局设置。
        首次配置时允许先保存代理再填写账号凭证。

        请求中的 access_token 会写入服务端配置文件，请仅从可信的管理端调用，
        不要把它放进浏览器端代码、日志或 URL。
      security:
        - basicAuth: []
        - appApiKey: []
        - appBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigUpdate'
            example:
              access_token: eyJhbGciOi...
              chatgpt_account_id: 00000000-0000-0000-0000-000000000000
              user_agent: codex-meter/1.0
              proxy_url: http://127.0.0.1:7890
              cache_ttl: 10m
              fedramp: false
      responses:
        '200':
          description: 更新后的非敏感配置
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/config/app-key:
    get:
      tags: [config]
      operationId: getAppAPIKey
      summary: 读取当前 App API Key
      description: |
        在已经通过 Basic Auth 或 App API Key 认证的情况下，返回当前维护的
        App API Key，供受保护的接口调试页面使用。响应设置为 no-store；请勿
        把返回值写入日志、URL 或前端源码。
      security:
        - basicAuth: []
        - appApiKey: []
        - appBearer: []
      responses:
        '200':
          description: 当前管理密钥
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppAPIKeyView'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/config/file:
    get:
      tags: [config]
      operationId: getConfigFile
      summary: 读取原始配置文件
      description: |
        返回服务端 config.json 的原文内容。该接口包含 OAuth Token、Cookie
        等敏感凭证，必须受 Basic Auth 或 App API Key 保护；在两者均未启用时
        服务端直接拒绝访问（403）。请勿把响应内容写入日志、URL 或前端源码。
      security:
        - basicAuth: []
        - appApiKey: []
        - appBearer: []
      responses:
        '200':
          description: 配置文件原文
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigFileView'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: 未启用任何管理认证，配置文件接口拒绝访问。
        '500':
          description: 服务端读取配置文件失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags: [config]
      operationId: updateConfigFile
      summary: 校验并保存原始配置文件
      description: |
        用请求体中的 content 整体替换服务端 config.json。内容必须是合法的
        配置 JSON，服务端会先写入临时文件完成完整性与凭证校验，再落盘并立即
        重新加载运行配置。修改 bind_addr 或 base_path 后需要重启服务才会生效。
        请仅从可信的管理端调用，不要把响应内容写入日志、URL 或前端源码。
      security:
        - basicAuth: []
        - appApiKey: []
        - appBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigFileUpdate'
            example:
              content: '{"cache_ttl":"10m"}'
      responses:
        '200':
          description: 保存后的配置文件原文
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigFileView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: 未启用任何管理认证，配置文件接口拒绝访问。
  /api/config/test:
    post:
      tags: [config]
      operationId: testConfig
      summary: 临时测试 OpenAI 额度连接
      description: |
        使用请求体中的临时配置与当前配置合并，请求一次 ChatGPT Web 的
        /backend-api/wham/usage。测试不会写入 config.json，也不会返回完整
        OAuth Token 或 Cookie。
      security:
        - basicAuth: []
        - appApiKey: []
        - appBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigUpdate'
      responses:
        '200':
          description: 测试成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigTestResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/UpstreamError'
...
  /audio:
    get:
      tags: [assets]
      operationId: getAudio
      summary: 获取内置提示音
      description: |
        返回嵌入在前端页面中的 WAV 提示音。kind 为 normal、warning 或 critical。
        此资源不需要 app API key，但如果服务启用了 Basic Auth，仍然需要 Basic Auth。
      security:
        - basicAuth: []
      parameters:
        - name: kind
          in: query
          required: true
          description: 提示音级别
          schema:
            type: string
            enum: [normal, warning, critical]
      responses:
        '200':
          description: WAV 音频
          content:
            audio/wav:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: 服务端启用 Basic Auth 时使用。
    appApiKey:
      type: apiKey
      in: header
      name: X-App-API-Key
      description: 配置 app_api_key 后使用的管理密钥。
    appBearer:
      type: http
      scheme: bearer
      bearerFormat: API key
      description: "与 X-App-API-Key 等价的 Authorization: Bearer 方式。"
  parameters:
    AccountID:
      name: account_id
      in: query
      required: false
      description: 本项目账号 id，省略或空字符串时固定使用主账号（优先 id=default，否则 accounts 的第一个账号）；页面展示切换不会改变该默认值。
      schema:
        type: string
...
    Force:
      name: force
      in: query
      required: false
      description: 是否绕过本地缓存直接请求上游。
      schema:
        type: boolean
        default: false
  responses:
    BadRequest:
      description: 请求参数或配置无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: force must be true or false
    Unauthorized:
      description: 未通过认证或未提供管理密钥
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: unauthorized
    UpstreamError:
      description: 上游额度、统计或预测服务请求失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    AccountConflict:
      description: 显式指定的账号凭证未配置完整，或尝试删除主账号、最后一个账号
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 资源不存在或 kind 无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    AccountSwitch:
      type: object
      required: [account_id]
      properties:
        account_id:
          type: string
      example:
        account_id: default
...